Adding missing specification for spikes_file and report file_name
Note that absolute path are not allowed. The rational is that if
instead of just filenames, absolute paths are also allowed, then
relative paths should also be allowed and that renders output_dir
pointless and/or ambiguous unless undesirable verbosity is added
to the spec to resolve the ambiguity.

Example simulation config fixed accordingly