Specify how to represent node types when `target_simulator` is "NEST" or "PyNN" (#33)
Removing table 2 about dynamics_param overrides.
This table has becomes confusing if we add non NEURON specific information
for point_neuron models. It also doesn't add anything that is not already said
above.

Removed the sentence about how NEURON interpret single sample point somas.
It could be confusing and it doesn't add information as the area of a sphere
and the revolution surface of a cylinder inscribing the sphere are equal.

Editing.